requirements.txt
python-telegram-bot[job-queue]
feedparser
requests
pytz
thefuzz
python-Levenshtein
beautifulsoup4
playwright
greenlet
